# backend/openapi.yaml
openapi: 3.1.0
info:
  title: School Management System API
  version: 1.0.0
servers:
  - url: http://localhost:3000
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Course:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        language:
          type: string
          enum: [English, Finnish, Swedish, Russian, German, French]
        level:
          type: string
          enum: [A1, A2, B1, B2, C1, C2]
        material:
          type: string
        instructorId:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        instructor:
          $ref: "#/components/schemas/PublicUser"
        sessions:
          type: array
          items:
            $ref: "#/components/schemas/ClassSessionSummary"
        enrollments:
          type: array
          items:
            $ref: "#/components/schemas/EnrollmentSummary"
    PublicUser:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
        secondName:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [student, instructor]
        comments:
          type: string
          nullable: true
    CourseSummary:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        language:
          type: string
          enum: [English, Finnish, Swedish, Russian, German, French]
        level:
          type: string
          enum: [A1, A2, B1, B2, C1, C2]
        material:
          type: string
        instructorId:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ClassSession:
      type: object
      properties:
        id:
          type: integer
        courseId:
          type: integer
        location:
          type: string
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
          nullable: true
        content:
          type: string
          nullable: true
        homework:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        course:
          $ref: "#/components/schemas/CourseSummary"
    ClassSessionSummary:
      type: object
      properties:
        id:
          type: integer
        courseId:
          type: integer
        location:
          type: string
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
          nullable: true
        content:
          type: string
          nullable: true
        homework:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    Enrollment:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        courseId:
          type: integer
        createdAt:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/PublicUser"
        course:
          $ref: "#/components/schemas/CourseSummary"
    EnrollmentSummary:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        courseId:
          type: integer
        createdAt:
          type: string
          format: date-time
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            message:
              type: string
            code:
              type: string
paths:
...
  /users/me/courses:
    get:
      summary: Get logged-in user courses
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Logged-in user courses
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Course"
        
        "401":
          description: Unauthorized

        "500":
          description: Server error
...
  /courses:
    get:
      summary: Get all courses
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of courses
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Course"
        
        "401":
          description: Unauthorized

        "500":
          description: Server error
    
    post:
      summary: Create course (instructor only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, language, level, material]
              properties:
                title:
                  type: string
                language:
                  type: string
                level:
                  type: string
                material:
                  type: string

      responses:
        "201":
          description: Course created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Course"

        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        
        "401":
          decription: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /courses/{id}:
    patch:
      summary: Update course information (instructor only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                language:
                  type: string
                level:
                  type: string
                material:
                  type: string
      
      responses:
        "200":
          description: Course updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Course"
        
        "400":
          decription: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        
        "401":
          decription: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

        "404":
          description: Course not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
       

    delete:
      summary: Delete course
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer

      responses:
        "200":
          description: Course deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      message:
                        type: string

        "400":
          description: Invalid ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        
        "401":
          decription: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /courses/{id}/sessions:
    get:
      summary: Get all sessions for a course
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of sessions
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/ClassSession"
        
        "401":
          description: Unauthorized

        "500":
          description: Server error
    
    post:
      summary: Create a new session for a course
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [location, startsAt]
              properties:
                courseId:
                  type: integer
                location:
                  type: string
                startsAt:
                  type: string
                  format: date-time
                endsAt:
                  type: string
                  format: date-time
                  nullable: true
      responses:
        "201":
          description: Session created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/ClassSession"

        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        
        "401":
          decription: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /courses/{id}/enroll:
    post:
      summary: Create enrollment (instructor only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId]
              properties:
                userId:
                  type: integer
      responses:
        "201":
          description: Enrollment created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Enrollment"

        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

        "409":
          description: Student already enrolled in this course
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...